02_MoodCatcher/requirements.txt
asgiref==3.2.3
astroid==2.3.3
attrs==19.3.0
backcall==0.1.0
beautifulsoup4==4.8.1
bleach==3.1.0
boto==2.49.0
boto3==1.10.32
botocore==1.13.32
bs4==0.0.1
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.1
cssselect==1.1.0
cycler==0.10.0
decorator==4.4.1
defusedxml==0.6.0
dj-database-url==0.5.0
Django==3.0
django-extensions==2.2.5
django-heroku==0.3.1
docutils==0.15.2
entrypoints==0.3
et-xmlfile==1.0.1
gensim==3.8.1
GetOldTweets3==0.0.11
gunicorn==20.0.4
idna==2.8
importlib-metadata==1.1.0
ipykernel==5.1.3
ipython==7.10.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
jdcal==1.4.1
jedi==0.15.1
Jinja2==2.10.3
jmespath==0.9.4
joblib==0.14.0
JPype1==0.7.0
json5==0.8.5
jsonschema==3.2.0

kiwisolver==1.1.0
konlpy==0.5.1
lazy-object-proxy==1.4.3
lxml==4.4.2
MarkupSafe==1.1.1
matplotlib==3.1.2
mccabe==0.6.1
mistune==0.8.4
more-itertools==8.0.0
nbconvert==5.6.1
nbformat==4.4.0
notebook==6.0.2
numpy==1.17.4
oauthlib==3.1.0
openpyxl==3.0.2
pandas==0.25.3
pandocfilters==1.4.2
parso==0.5.1
pickleshare==0.7.5
prometheus-client==0.7.1
prompt-toolkit==3.0.2
psycopg2==2.8.4
Pygments==2.5.2
pylint==2.4.4
pyparsing==2.4.5
pyquery==1.4.1
pyrsistent==0.15.6
PySocks==1.7.1
python-dateutil==2.8.0
python-decouple==3.3
pytz==2019.3
pywinpty==0.5.7
pyzmq==18.1.1
requests==2.22.0
requests-oauthlib==1.3.0
s3transfer==0.2.1
scikit-learn==0.22
scipy==1.3.3
Send2Trash==1.5.0
six==1.13.0
smart-open==1.9.0
soupsieve==1.9.5
sqlparse==0.3.0
terminado==0.8.3
testpath==0.4.4
tornado==6.0.3
tqdm==4.39.0
traitlets==4.3.3
tweepy==3.8.0
typed-ast==1.4.0
urllib3==1.25.7
virtualenv==16.7.8
wcwidth==0.1.7
webencodings==0.5.1
whitenoise==5.0.1
widgetsnbextension==3.5.1
wrapt==1.11.2
xlrd==1.2.0
zipp==0.6.0
